requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
Flask==2.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
py-healthcheck==1.10.1
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.4
requests==2.28.1
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.14.1
